+++ custom_components/omniremote/const.py
@@ -7,7 +7,7 @@ from typing import Any
 import uuid
 
 DOMAIN = "omniremote"
-VERSION = "1.9.8"
+VERSION = "1.10.0"
 
 # Storage
 STORAGE_VERSION = 1